application context + database access rewirte
